updated to remove references to notes in code
