+++ lib/news/non_readable_news_post.dart
@@ -33,11 +33,11 @@ import 'package:communitypod/constants/app.dart';
 import 'package:communitypod/constants/colours.dart';
 import 'package:communitypod/constants/ui.dart';
 import 'package:communitypod/models/news.dart';
-import 'package:communitypod/news/list_notes_screen.dart';
+import 'package:communitypod/news/list_news_screen.dart';
 import 'package:communitypod/news/share_news.dart';
 import 'package:communitypod/widgets/msg_card.dart';
-import 'package:communitypod/widgets/note_action_button.dart';
-import 'package:communitypod/widgets/note_display_metadata.dart';
+import 'package:communitypod/widgets/action_button.dart';
+import 'package:communitypod/widgets/display_metadata.dart';
 
 /// A [stateful] widget for displaying a message when the user tries to view
 /// an externally owned news file.
@@ -96,7 +96,7 @@ class _NonReadableNewsPostState extends State<NonReadableNewsPost> {
         child: Column(
           children: <Widget>[
             // Display note metadata - show sharing and path info but not dates (as requires newsContent)
-            DisplayNewsMetadata(
+            DisplayMetadata(
               newsOwner: _newsPost.newsOwner,
               permissionGranter: _newsPost.permissionGranter!,
               permissionList: _newsPost.permissionList,
