+++ lib/widgets/custom_back_button.dart
@@ -40,7 +40,7 @@ import 'package:communitypod/widgets/save_dialog.dart';
 /// Arguments:
 /// - [childPage] - The child page to navigate back to.
 ///   [scaffoldController] - Controller for the Solid scaffold.
-/// - [textController] - Optional text controller if back is being called from note editor.
+/// - [textController] - Optional text controller if back is being called from news editor.
 /// - [formKey] - Key of the form to edit news post metadata
 /// - [prevNews] - Optional existing user's news data object. Required
 /// for saving existing news posts. (Default: null).
@@ -49,8 +49,8 @@ import 'package:communitypod/widgets/save_dialog.dart';
 /// - [isExisting] - Optional boolean denoting whether news post already
 /// exists. (Default: false).
 
-class NewsBackButton extends StatelessWidget {
-  const NewsBackButton({
+class CustomBackButton extends StatelessWidget {
+  const CustomBackButton({
     super.key,
     required this.childPage,
     required this.scaffoldController,
@@ -82,7 +82,7 @@ class NewsBackButton extends StatelessWidget {
       onPressed: () {
         if (formKey?.currentState?.saveAndValidate() ?? false) {
           if (textController != null) {
-            String noteText = textController!.text;
+            String newsText = textController!.text;
             Map formData = formKey?.currentState?.value as Map;
             String newsTitle = formData[newsTitlePred].replaceAll('\n', '');
 
@@ -91,7 +91,7 @@ class NewsBackButton extends StatelessWidget {
               prevNewsTitle = prevNews!.content!.newsTitle;
               prevNewsContent = prevNews!.content!.newsContent;
               // Check if title or content changed
-              if (newsTitle != prevNewsTitle || noteText != prevNewsContent) {
+              if (newsTitle != prevNewsTitle || newsText != prevNewsContent) {
                 showDialog<void>(
                   context: context,
                   barrierDismissible: false, // user must tap button!
