+++ lib/widgets/edit_scroll_view.dart
@@ -1,4 +1,4 @@
-/// A widget for creating and editing notes in a SingleChildScrollView()
+/// A widget for creating and editing news posts in a SingleChildScrollView()
 ///
 // Time-stamp: <Wednesday 2025-07-18 16:17:37 +1000 Jess Moore>
 ///
@@ -36,15 +36,27 @@ import 'package:communitypod/constants/app.dart';
 import 'package:communitypod/constants/turtle_structures.dart';
 import 'package:communitypod/models/news.dart';
 import 'package:communitypod/widgets/markdown_editor.dart';
-import 'package:communitypod/widgets/note_back_button.dart';
-import 'package:communitypod/widgets/note_save_button.dart';
+import 'package:communitypod/widgets/custom_back_button.dart';
+import 'package:communitypod/widgets/save_button.dart';
 
 /// A [StatelessWidget] widget setup for calling the
-/// SingleChildScrollView() to edit a note, whether a
-/// new note or a pre-existing note, and whether owned
-/// by the user or shared to the user.
-class NewsEditScrollView extends StatelessWidget {
-  const NewsEditScrollView({
+/// SingleChildScrollView() to edit a news post.
+///
+/// Arguments:
+/// - [formKey] - key identifier for news post form.
+/// - [textController] - text in form.
+/// - [scrollController] - controller for scroll widget.
+/// - [scaffoldController] - controller for scaffold widget.
+/// - [focusTitle] - [FocusNode] for title form field.
+/// - [focusContent] - [FocusNode] for text content form field.
+/// - [childPage] - page to navigate to on back button press.
+/// - [data] - data from text controller.
+/// - [prevNews] - content of earlier news post version if post not new.
+/// - [isExternal] - flag describing whether file externally owned.
+/// - [isExisting] - flag describing whether earlier version of file exists.
+/// - [newsTitle] - the title of earlier news post version if post not new.
+class EditScrollView extends StatelessWidget {
+  const EditScrollView({
     super.key,
     required this.formKey,
     required TextEditingController? textController,
@@ -73,28 +85,28 @@ class NewsEditScrollView extends StatelessWidget {
   /// Scaffold controller
   final SolidScaffoldController _scaffoldController;
 
-  /// Focus node for note title field
+  /// Focus node for title field
   final FocusNode _focusTitle;
 
-  /// Focus node for note contents text field
+  /// Focus node for contents text field
   final FocusNode _focusContent;
 
   /// Childpage used by back button
   final Widget childPage;
 
+  /// Text field data
   final String data;
 
-  /// Existing note data is note already exists
+  /// Existing data if file already exists
   final News? prevNews;
 
-  /// Boolean describing whether note is shared to pod owner from an
-  /// external source.
+  /// Boolean describing whether file is externally owned.
   final bool isExternal;
 
-  /// Boolean describing whether note already exists.
+  /// Boolean describing whether file already exists.
   final bool isExisting;
 
-  /// Title of note where note already exists
+  /// Data title field if file already exists
   final String? newsTitle;
 
   @override
@@ -102,7 +114,7 @@ class NewsEditScrollView extends StatelessWidget {
     String currDateStr = '';
 
     if (!isExisting) {
-      // New note: fetch current date for heading
+      // New file: fetch current date for heading
       currDateStr =
           DateFormat('dd MMMM yyyy').format(DateTime.now()).toString();
     }
@@ -132,8 +144,8 @@ class NewsEditScrollView extends StatelessWidget {
                   isExternal: isExternal,
                 ),
                 // Back button
-                // Nav to view note or view isExternal note
-                NewsBackButton(
+                // Nav to child page
+                CustomBackButton(
                   childPage: childPage,
                   textController: _textController,
                   formKey: formKey,
@@ -175,7 +187,7 @@ class NewsEditScrollView extends StatelessWidget {
                       child: Column(
                         spacing: 10.0,
                         children: [
-                          // New note: show current date
+                          // New file: show current date
                           if (!isExisting) ...[
                             Row(
                               crossAxisAlignment: CrossAxisAlignment.start,
@@ -187,8 +199,8 @@ class NewsEditScrollView extends StatelessWidget {
                               ],
                             ),
                           ],
-                          // Edit existing note: populated text field with
-                          // previous note data
+                          // Edit existing file: populated text field with
+                          // previous data
                           FormBuilderTextField(
                             name: newsTitlePred,
                             initialValue: (isExisting) ? newsTitle : null,
