+++ lib/widgets/highlight_image.dart
