+++ lib/widgets/item_trailing_buttons.dart
@@ -1,4 +1,4 @@
-/// A stateless widget to show trailing buttons in a note list item.
+/// A stateless widget to show trailing buttons in a list item.
 ///
 /// Copyright (C) 2026 Software Innovation Institute, Australian National University
 ///
@@ -27,31 +27,30 @@ import 'package:flutter/material.dart';
 import 'package:solidui/solidui.dart';
 
 import 'package:communitypod/models/news.dart';
-import 'package:communitypod/news/list_notes_screen.dart';
+import 'package:communitypod/news/list_news_screen.dart';
 import 'package:communitypod/news/share_news.dart';
 import 'package:communitypod/widgets/simple_action_button.dart';
 
-/// A [stateless] widget to show trailing buttons in a note
-/// list item.
+/// A [stateless] widget to show trailing buttons in a list item.
 ///
 /// Arguments:
-/// - [note] - A note.
+/// - [item] - A list item.
 /// - [scaffoldController] - Controller for the Solid scaffold.
 ///
 class ItemTrailingButtons extends StatelessWidget {
   const ItemTrailingButtons({
     super.key,
-    required News note,
+    required News item,
     required SolidScaffoldController scaffoldController,
-  })  : _note = note,
+  })  : _item = item,
         _scaffoldController = scaffoldController;
 
-  final News _note;
+  final News _item;
   final SolidScaffoldController _scaffoldController;
 
   @override
   Widget build(BuildContext context) {
-    List accessList = _note.permissionList.split(',');
+    List accessList = _item.permissionList.split(',');
 
     return Row(
       mainAxisAlignment: MainAxisAlignment.end,
@@ -62,9 +61,9 @@ class ItemTrailingButtons extends StatelessWidget {
           SimpleActionButton(
             icon: const Icon(Icons.share),
             childPage: ShareNews(
-              newsUrl: _note.newsUrl,
-              newsOwner: _note.newsOwner,
-              isExternal: _note.isExternalRes,
+              newsUrl: _item.newsUrl,
+              newsOwner: _item.newsOwner,
+              isExternal: _item.isExternalRes,
               backPage: ListNewsScreen(
                 scaffoldController: _scaffoldController,
               ),
@@ -73,8 +72,7 @@ class ItemTrailingButtons extends StatelessWidget {
             scaffoldController: _scaffoldController,
           ),
         ],
-        // Open note icon
-        // Launch icon to denote expanding to full size
+        // Launch icon to denote viewing full post
         // Alternative open_in_full_rounded
         const Icon(Icons.launch_rounded),
       ],
